refactor: update LDAP sync tests to use MikroORM and streamline module imports
