+++ apps/server/src/modules/ldap-sync-console/api/ldap-sync.console.spec.ts
@@ -1,58 +1,43 @@
-import { createMock, DeepMocked } from '@golevelup/ts-jest';
+import { MikroORM } from '@mikro-orm/core';
 import { Test, TestingModule } from '@nestjs/testing';
-import { ConsoleWriterService } from '@infra/console';
 
-// Mock the UC module to avoid loading legacy Feathers code which has problematic imports
-jest.mock('./ldap-sync.uc', () => {
+const mockRunLegacyLdapSync = jest.fn();
+jest.mock('@imports-from-feathers', () => {
 	return {
-		LdapSyncUc: jest.fn().mockImplementation(() => {
-			return {
-				runLdapSync: jest.fn(),
-			};
-		}),
+		runLegacyLdapSync: mockRunLegacyLdapSync,
 	};
 });
 
+import { ConsoleWriterService } from '@infra/console';
+import { LdapSyncConsoleAppTestModule } from '../ldap-sync-console.app.module';
 import { LdapSyncConsole } from './ldap-sync.console';
-import { LdapSyncUc } from './ldap-sync.uc';
+// Mock process.exit to prevent test from exiting
+const mockExit = jest.spyOn(process, 'exit').mockImplementation(() => undefined as never);
 
 describe(LdapSyncConsole.name, () => {
 	let module: TestingModule;
-	let console: LdapSyncConsole;
-	let ldapSyncUc: DeepMocked<LdapSyncUc>;
-	let consoleWriter: DeepMocked<ConsoleWriterService>;
+	let consoleService: LdapSyncConsole;
+	let orm: MikroORM;
+	let consoleWriter: ConsoleWriterService;
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
-			providers: [
-				LdapSyncConsole,
-				{
-					provide: LdapSyncUc,
-					useValue: createMock<LdapSyncUc>(),
-				},
-				{
-					provide: ConsoleWriterService,
-					useValue: createMock<ConsoleWriterService>(),
-				},
-			],
+			imports: [LdapSyncConsoleAppTestModule],
 		}).compile();
 
-		console = module.get(LdapSyncConsole);
-		ldapSyncUc = module.get(LdapSyncUc);
+		consoleService = module.get(LdapSyncConsole);
+		orm = module.get(MikroORM);
 		consoleWriter = module.get(ConsoleWriterService);
 	});
 
-	beforeEach(() => {
-		jest.clearAllMocks();
-	});
-
 	afterAll(async () => {
+		mockExit.mockRestore();
 		await module.close();
 	});
 
 	describe('when console is initialized', () => {
 		it('should be defined', () => {
-			expect(console).toBeDefined();
+			expect(consoleService).toBeDefined();
 		});
 	});
 
@@ -61,7 +46,7 @@ describe(LdapSyncConsole.name, () => {
 			const setup = () => {
 				const options = { deltaSync: false };
 				const stats = { success: true, errors: [], systems: {} };
-				ldapSyncUc.runLdapSync.mockResolvedValueOnce(stats);
+				mockRunLegacyLdapSync.mockResolvedValueOnce(stats);
 
 				return { options, stats };
 			};
@@ -69,83 +54,20 @@ describe(LdapSyncConsole.name, () => {
 			it('should call runLdapSync with forceFullSync=true', async () => {
 				const { options } = setup();
 
-				await console.sync(options);
+				await consoleService.sync(options);
 
-				expect(ldapSyncUc.runLdapSync).toHaveBeenCalledWith(true);
+				expect(mockRunLegacyLdapSync).toHaveBeenCalledWith(orm, { forceFullSync: true });
 			});
 
 			it('should log start and finish messages', async () => {
 				const { options } = setup();
+				const spy = jest.spyOn(consoleWriter, 'info');
 
-				await console.sync(options);
+				await consoleService.sync(options);
 
-				expect(consoleWriter.info).toHaveBeenCalledTimes(2);
-				expect(consoleWriter.info).toHaveBeenNthCalledWith(
-					1,
-					expect.stringContaining('"message":"Starting LDAP synchronization"')
-				);
-				expect(consoleWriter.info).toHaveBeenNthCalledWith(
-					2,
-					expect.stringContaining('"message":"LDAP synchronization finished"')
-				);
-			});
-		});
-
-		describe('when called with deltaSync=true', () => {
-			const setup = () => {
-				const options = { deltaSync: true };
-				const stats = { success: true, errors: [], systems: {} };
-				ldapSyncUc.runLdapSync.mockResolvedValueOnce(stats);
-
-				return { options, stats };
-			};
-
-			it('should call runLdapSync with forceFullSync=false', async () => {
-				const { options } = setup();
-
-				await console.sync(options);
-
-				expect(ldapSyncUc.runLdapSync).toHaveBeenCalledWith(false);
-			});
-		});
-
-		describe('when called with deltaSync="true" as string', () => {
-			const setup = () => {
-				const options = { deltaSync: 'true' as unknown as boolean };
-				const stats = { success: true, errors: [], systems: {} };
-				ldapSyncUc.runLdapSync.mockResolvedValueOnce(stats);
-
-				return { options, stats };
-			};
-
-			it('should call runLdapSync with forceFullSync=false', async () => {
-				const { options } = setup();
-
-				await console.sync(options);
-
-				expect(ldapSyncUc.runLdapSync).toHaveBeenCalledWith(false);
-			});
-		});
-
-		describe('when sync returns errors', () => {
-			const setup = () => {
-				const options = { deltaSync: false };
-				const stats = {
-					success: false,
-					errors: ['Error 1', 'Error 2'],
-					systems: { ldap1: { success: false } },
-				};
-				ldapSyncUc.runLdapSync.mockResolvedValueOnce(stats);
-
-				return { options, stats };
-			};
-
-			it('should log the stats including errors', async () => {
-				const { options, stats } = setup();
-
-				await console.sync(options);
-
-				expect(consoleWriter.info).toHaveBeenNthCalledWith(2, expect.stringContaining(JSON.stringify(stats)));
+				expect(spy).toHaveBeenCalledTimes(2);
+				expect(spy).toHaveBeenNthCalledWith(1, expect.stringContaining('"message":"Starting LDAP synchronization"'));
+				expect(spy).toHaveBeenNthCalledWith(2, expect.stringContaining('"message":"LDAP synchronization finished"'));
 			});
 		});
 	});
